+++ Sources/SuperwallKit/Misc/Extensions/Publishers/Publisher+HasRetrieved.swift
@@ -12,7 +12,13 @@ extension Publisher {
   func hasValue<T>() async -> T where Output == T? {
     await self
       .compactMap { $0 }
-      .eraseToAnyPublisher()
       .async()
+  }
+
+  @discardableResult
+  func throwableHasValue<T>() async throws -> T where Output == T? {
+    try await self
+      .compactMap { $0 }
+      .throwableAsync()
   }
 }