+++ Sources/SuperwallKit/Paywall/Presentation/Internal Presentation/Presentation Request/PresentationRequest.swift
@@ -9,9 +9,18 @@ import UIKit
 import Combine
 
 enum PresentationRequestType: Equatable, CustomStringConvertible {
+  /// Presenting via ``Superwall/register(event:params:handler:feature:)``.
   case presentation
+
+  /// Getting the paywall view controller via
+  /// ``Superwall/getPaywall(forEvent:params:paywallOverrides:delegate:)``.
   case getPaywallViewController(PaywallViewControllerDelegateAdapter)
+
+  /// Getting the presentation result via ``Superwall/getPresentationResult(forEvent:)``
   case getPresentationResult
+
+  /// Getting the presentation result from an event that's used internally. Specifically, getting the result
+  /// of calling `paywall_decline`.
   case getImplicitPresentationResult
 
   var description: String {
@@ -74,7 +83,6 @@ struct PresentationRequest {
     var subscriptionStatus: AnyPublisher<SubscriptionStatus, Never>
     var isPaywallPresented: Bool
     var type: PresentationRequestType
-    var hasInternet: Bool
   }
   var flags: Flags