+++ Tests/SuperwallKitTests/Paywall/Presentation/Internal Presentation/Operators/CheckPaywallPresentableOperatorTests.swift
@@ -42,6 +42,7 @@ final class CheckPaywallPresentableOperatorTests: XCTestCase {
       .setting(\.flags.subscriptionStatus, to: publisher)
 
     let paywallVc = dependencyContainer.makePaywallViewController(for: .stub(), withCache: nil, delegate: nil)
+    paywallVc.loadViewIfNeeded()
     let expectation = expectation(description: "Called publisher")
     do {
       try await Superwall.shared.getPresenter(
@@ -95,11 +96,12 @@ final class CheckPaywallPresentableOperatorTests: XCTestCase {
       isDebuggerLaunched: false,
       subscriptionStatus: inactiveSubscriptionPublisher,
       isPaywallPresented: false,
-      type: .getPaywallViewController(.stub())
+      type: .presentation
     )
     .setting(\.presenter, to: nil)
 
     let paywallVc = dependencyContainer.makePaywallViewController(for: .stub(), withCache: nil, delegate: nil)
+    paywallVc.loadViewIfNeeded()
     let expectation = expectation(description: "Called publisher")
     do {
       try await Superwall.shared.getPresenter(
@@ -143,6 +145,7 @@ final class CheckPaywallPresentableOperatorTests: XCTestCase {
 
     let dependencyContainer = DependencyContainer()
     let paywallVc = dependencyContainer.makePaywallViewController(for: .stub(), withCache: nil, delegate: nil)
+    paywallVc.loadViewIfNeeded()
     do {
       try await Superwall.shared.getPresenter(
         for: paywallVc,